Keep model file handles in ScreenAIServiceImpl.
To support lazy-load for OCR service, ScreenAI service will keep the
model files' handles as long as the service is running, so that OCR
would be able to ask for model files later as well.
This also will help to support the possibility of unloading OCR and
reloading it later when it's needed.

AX-Relnotes: n/a
Bug: 354063627
Change-Id: Iba0a56bd28dbbc4cb30f49ab12309906bd45d9ea
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5731893
Auto-Submit: Ramin Halavati <rhalavati@chromium.org>
Commit-Queue: Ramin Halavati <rhalavati@chromium.org>
Reviewed-by: David Tseng <dtseng@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1334103}