[views-ax] Migrates kDefaultActionVerb attribute to new system in combobox
This CL migrates the kDefaultActionVerb to whenever its value should
change, rather than querying the value and computing it only when needed
in the following files:
 1. ui\views\controls\combobox\combobox.cc
 2. ui\views\controls\editable_combobox\editable_combobox.cc

This CL eliminates the direct assignment of the kDefaultActionVerb
attribute within the View::GetAccessibleNodeData method. Instead, it
introduces a mechanism to update the kDefaultActionVerb attribute in the
accessibility cache when any attribute that contributes to the
accessible kDefaultActionVerb undergoes a change. This approach ensures
that the accessible kDefaultActionVerb remains up-to-date in the
accessibility cache.

This CL is part of the ViewsAX project:
https://docs.google.com/document/d/1Ku7HOyDsiZem1yaV6ccZ-tz3lO2XR2NEcm8HjR6d-VY/edit#heading=h.ke1u3utej413

Bug: 325137417
Change-Id: Iffd66ba770b648a69f87cd2da7cb6371157b849f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5705456
Reviewed-by: Javier Contreras <javiercon@microsoft.com>
Reviewed-by: Dirk Pranke <dpranke@google.com>
Reviewed-by: Benjamin Beaudry <benjamin.beaudry@microsoft.com>
Commit-Queue: Ashish Kumar <ashishkum@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1334112}