Support HTTP/2 IP-based pooling in HttpStreamPool
Each time service endpoint results change, check to see if there is an
active SPDY session that can be used for requests based on resolved IP addresses.

Add a method to SpdySessionPool to perform the check. The method is
similar to SpdySessionPool::OnHostResolutionComplete(), but is much
simpler because there should only be one DNS resolution in the
HttpStreamPool at a time.

Bug: 346835898
Change-Id: I57deca3a9de7ac836b2c36df5253ce4dad1e4cc2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5725437
Reviewed-by: Adam Rice <ricea@chromium.org>
Commit-Queue: Kenichi Ishibashi <bashi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1334121}