[UNOp2][Android] Make bottom sheet progress view height match shown view
In the sign-in bottom sheet, the progress state view is often shown
for a very short duration. Since it has a fixed height, the bottom
sheet "jumps" visually when showing it, which is disturbing especially
during the managed account sign-in flow given the progress state needs
to be shown twice.

This CL makes the progress view have the same height as the previous
state's view, just before showing the progress view.

See https://crbug.com/327127097#comment4 for recordings.

Bug: 327127097
Change-Id: I99e728502d72a36f43eae679574f1c1dc3784390
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5689141
Commit-Queue: Menghan Yang <myuu@google.com>
Reviewed-by: Boris Sazonov <bsazonov@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1334128}