Add model type controller for PLUS_ADDRESS(_SETTING)
Adds a common model type controller for the PLUS_ADDRESS and
PLUS_ADDRESS_SETTING types. This is motivated by the fact that the types
should be disabled for Dasher users - a precondition that will be added
in a follow-up CL.

Existing sync code for the two model types lives in
components/plus_addresses/webdata and components/plus_addresses/settings
respectively. For this reason, the common model type controller is
placed in a new components/plus_addresses/sync_utils directory.

This CL takes the opportunity to move the feature check for
kPlusAddressesEnabled into the model type controller. Why this might be
desirable is discussed at [1].

[1] https://chromium-review.googlesource.com/c/chromium/src/+/5717930/comment/10ba066b_4d8d82f9/

Bug: 354688446
Change-Id: Ib339bc475c9bc7dabbf4ce102e8b1b509c64c412
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5730857
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Jan Keitel <jkeitel@google.com>
Commit-Queue: Florian Leimgruber <fleimgruber@google.com>
Reviewed-by: Rushan Suleymanov <rushans@google.com>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1334132}