[Floating SSO] Observe cookie changes and commit them to store
Fetch the list of already existing cookies (in case cookies were
created before the service starts listening) and listen for cookie
changes.

Filter cookies based on the requirements in http://go/floating-sso-dd.
If the cookie passes the filter, it is added to the cookie store in
the bridge. Delete the cookie from the store if the cookie change
cause is the cookie being deleted, overwritten, expired.

Bug: b:346354255
Change-Id: I08657a5a79f6c3e78755d2999e7854e9448b3e45
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5724694
Reviewed-by: Marc Treib <treib@chromium.org>
Reviewed-by: Andrey Davydov <andreydav@google.com>
Commit-Queue: Maria Petrisor <mpetrisor@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1334212}