[skyvault] Make NotificationManager a KeyedService
Changes the MigrationNotificationManager to a KeyedService, so that the
same instance can be used from different points in the process, e.g. to
show that the migration is starting (high-level in MigrationManager),
but also to prompt the user to sign-in to OneDrive if needed (deep in
the upload flow).

R=poromov@chromium.org

Bug: b/355183821
Change-Id: I84db15b84ef27a5f8568542d318844a7d7e1cc4d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5741675
Reviewed-by: Sergey Poromov <poromov@chromium.org>
Commit-Queue: Aida Zolic <aidazolic@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1334215}