Force supervised user reauthentication for YouTube with interstitial
When a supervised user in the pending state tries to access YouTube,
the re-auth interstitial will be displayed in the same tab.
Completing this re-auth will redirect the user to YouTube.

This interstitial is only available on Desktop platforms and will not
be displayed in sub-frames.

Low-Coverage-Reason: TESTS_IN_SEPARATE_CL
Bug: 346264927
Change-Id: I41179a5f4e5d068eebf31e53908f669fda18393d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5730275
Reviewed-by: Denis Kuznetsov <antrim@chromium.org>
Reviewed-by: James Lee <ljjlee@google.com>
Commit-Queue: Duong Dac <ddac@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1334220}