[iOS] Remove TabGridPaging protocol
Implement the properties directly in the classes as it was never used
as a protocol elsewhere in the app.

Also make sure to set the TabGridMode and the ActivePage through
mediators.

Bug: None
Change-Id: I5dbb67f6cf3425a220169d8ed35db6594f58d4db
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5743502
Reviewed-by: Aliona Dangla <alionadangla@chromium.org>
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Auto-Submit: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1334280}