AW: ensure URLLoaderThrottle is always used on UI thread
This URLLoaderThrottle is a single-threaded class which is only safe to
call on the UI thread. This is generally upheld by code which calls
URLLoaderThrottle override methods, however there was one private method
which was being called from Java. It's possible that the Java
implementation code could call this on the Main Looper instead of the
Chromium UI thread, so we can solve this by posting to the UI thread
instead.

This also introduces a utility class (BackgroundThreadExecutor) which
I'm using for testing, however I intend to also use the same class in
the production code in a future change.

Fixed: 355541498
Test: run_webview_instrumentation_test_apk -f AwSupervisedUserTest.*
Change-Id: I8bc13fff95c337893cd4bebb060c8b9ccf3d70ae
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5742930
Reviewed-by: Aashna Jena <aashnajena@google.com>
Auto-Submit: Nate Fischer <ntfschr@chromium.org>
Commit-Queue: Aashna Jena <aashnajena@google.com>
Cr-Commit-Position: refs/heads/main@{#1334301}