[PDF Ink Signatures] Enable undo/redo even when not in annotation mode
With the new UI refresh, the undo and redo buttons are now in the viewer
toolbar, which is always displayed, even when the viewer is not in
annotation mode.

Since the buttons can always be clicked, allow the backend to undo/redo
even when annotation mode is disabled.

Bug: 355578909
Change-Id: I5c8fc11c3882b8e0f536b9bd9b182372d039678b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5744321
Reviewed-by: Lei Zhang <thestig@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Alan Screen <awscreen@chromium.org>
Commit-Queue: Andy Phan <andyphan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1334411}