[Oidc Enrollment] BrowserSignin also disables Dasher-based OIDC enrollment
The last step of Dasher-based OIDC enrollment involves setting the new account as the primary account of the new ODIC profile. However, this causes a crash when `BrowserSignin` is set to 0 (aka Disabled).

Add a bypass logic where Dasher-based OIDC enrollment is interrupted right after registration if `BrowserSignin` is set to 0, to just prevent crash for now.

The interaction between the policy and OIDC enrollment will be finalized later and solved in 355270189.

Bug: 355272202
Change-Id: I844848d96ea9645160e377281e3c8ac72d1ad1b3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5743997
Commit-Queue: Zonghan Xu <xzonghan@chromium.org>
Reviewed-by: Owen Min <zmin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1334418}