Implement Sanitize initial dialog in SWA
This CL implements the sanitize initial dialog. Currently this UI
doesn't do anything, none of the buttons are connected to the
functionalities that they should perform. Those will be added in a
separate CL in the interest of a more reviewable CL.
The way the UI currently looks is as follows:
https://screenshot.googleplex.com/8ps3fpBv54Riw3L

BUG=b:332697577, b:332698769
TEST=Manual opening of SWA

Change-Id: I46104e33fa5f083cc4225e8dc1db3680da78650a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5645977
Reviewed-by: John Admanski <jadmanski@chromium.org>
Commit-Queue: Behnood Momenzadeh <behnoodm@chromium.org>
Reviewed-by: Jimmy Gong <jimmyxgong@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1334422}