Add MatchDirection method to gfx::Range and gfx::RangeF
This CL adds a simple method to the gfx::Range classes that
will match the direction of the provided range, along with
a basic test. It also updates a few places in render_text.cc to
use this new method.

The primary motivation for this CL is a larger CL it was split from:
https://chromium-review.googlesource.com/c/chromium/src/+/5634309

Bug: 40270295, 40267908
Change-Id: I73a8f862a09d20e01cdab5696b6bd69981b73bd3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5742886
Commit-Queue: Kurt Catti-Schmidt <kschmi@microsoft.com>
Reviewed-by: Avi Drissman <avi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1334439}