Rename WasFocused to WasFocusedByTapOrClick in Autofill2.FieldInfo UKM event
The current WasFocused metric does not return whether a field was
focused. It returns whether a field was clicked or tapped. If the user
focuses a field via the tab-key, autofill is not queried for
suggestions and the bit is not set. In this CL, we rename WasFocused
to WasFocusedByTapOrClick, and set WasFocused from Autofill's
|was_focused| value in Autofill2.FieldInfo UKM event.

Bug: 1325851
Change-Id: Ief3843cff24d7354bc5f224dd24c0aed15d1c834
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5726355
Reviewed-by: Dominic Battré <battre@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Lan Wei <lanwei@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1334459}