[WebView Shell] Make WebView Shell able to use net_java library
In order to be able to import the net_java library into the WebView
shell APK, a few minSDK errors need to be rectified so that the minSDK
of 21 does not conflict with the minSDK of a couple of classes in the
net_java library, which require minSDK of 23 & 24 or greater
(Android M and N).
See crrev/c/5736379 for the relevant findings.

Bug: b:355497433
Test: See linked bug for test details
Change-Id: Ife742202d733e9037935cc721921c5a6cc5368d2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5741944
Reviewed-by: Adam Langley <agl@chromium.org>
Commit-Queue: Adam Walls <avvall@chromium.org>
Reviewed-by: Nate Fischer <ntfschr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1334480}