updater: restore `brand` from/to registry, and `brand` is permanent
If a `brand` code already exists in the registry in `ClientState`, this
brand code is considered authoritative, and overrides any brand code
that is recorded in `prefs`.

If a `brand` code does not exist in the registry, the value in `prefs`
is written to the registry.

If a `brand` code already exists in the registry or the prefs for an
app, the app is considered "branded" to that brand, and no new brand can
be set for that app.

Fixed: 355693627
Change-Id: I035aeebc8114e96c470035127f1190b3fd333a88
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5747313
Auto-Submit: S Ganesh <ganesh@chromium.org>
Reviewed-by: Joshua Pawlicki <waffles@chromium.org>
Commit-Queue: Joshua Pawlicki <waffles@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: S Ganesh <ganesh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1334483}