Reland "Add background blur to the AutoPiP scrim"
This is a reland of https://crrev.com/c/5661379. That change was
reverted, because the blur had no effect on Mac.

The fix involved creating a new view for the blur effect,  and adding
it as a child of the `AutoPipSettingOverlayView`. Instead of adding the
blur to the background layer.

Bug: 349627097
Change-Id: I1240fb29353378e02ff9a157c1e500d1417452d2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5742627
Commit-Queue: Benjamin Keen <bkeen@google.com>
Reviewed-by: Frank Liberato <liberato@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1334507}