[infra] Ensure a sane ordering for link nodes.
Currently the order of children or parents of a link node is the default
which is in key order. This would be expected to be the same as the
order that the nodes were created except that the unique portion of the
keys is the string representation of an incrementing numeric value. This
means that keys with different orders of magnitude will not necessarily
be in the expected order (e.g. "10" < "9"). To ensure a consistent
reasonable order, use DEFINITION_ORDER instead which returns the parent
or child nodes in the order that the edges were created.

Change-Id: I11ae5c5ca6621bb4f63159b9c758b20aecc7af21
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5745843
Commit-Queue: Struan Shrimpton <sshrimp@google.com>
Auto-Submit: Garrett Beaty <gbeaty@google.com>
Reviewed-by: Struan Shrimpton <sshrimp@google.com>
Cr-Commit-Position: refs/heads/main@{#1334519}