webui: fix wrong auto-resized host size on high DPI screen
The WebUI host is resized to RenderFrameHost::GetFrameSize() but it is
not device-independent, causing the bubble size to be 2x the actual size
on mac built-in screen. This patch scales the frame size to the device
independent size.

Bug: 355554869
Change-Id: I5d15f591e62f079463735b1fcc05edd59dfce113
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5745179
Reviewed-by: Thomas Lukaszewicz <tluk@chromium.org>
Commit-Queue: Keren Zhu <kerenzhu@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1334587}