Prerender: Add kOtherPrerenderedPageActivated final status
When one of prerendered pages is reserved for activation, other
prerendered pages are destroyed with kTriggerDestroyed. However,
kTriggerDestroyed is used in multiple places, so it is not so useful to
identify the exact reason of the destruction. To improve it, this CL
introduces kOtherPrerenderedPageActivated.

NO_IFTTT=Necessary changes will be done on a separate repository

Bug: 40250648, 356049753
Change-Id: I1469cbf7391eaa629b13ba23c563f99c91aad8da
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5741048
Reviewed-by: Taiyo Mizuhashi <taiyo@chromium.org>
Commit-Queue: Hiroki Nakagawa <nhiroki@chromium.org>
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1334637}