Modularize feedback
This CL is a refactor with no intended behavior change.

The desktop feedback code is split into two targets: feedback (the
interface) and feedback_impl (the implementation). This is to work
around the many components that both need to invoke the feedback dialog
(the interface), but which also must provide state to the feedback
upload implementation.

Bug: 353332589
Change-Id: I644dd62869b163448c45c964e74971424e15d7e7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5747242
Reviewed-by: Thomas Lukaszewicz <tluk@chromium.org>
Commit-Queue: Erik Chen <erikchen@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1334664}