Extensions: Ensure permissions prompt is shown on the correct window
When searching a window for a given ExtensionFunction, WindowControllerList
looks for the focused window. If no window is focused, it falls back to
the last window in the list.

This behavior can be problematic if the extension API call originates
from an extension popup. Extension popups are hosted in a window which
is different from the browser window. The browser window is not in a
focused state while the popup is active. Thus, the search always returns
the last window in the list. This can lead to a wrong window selection
in cases when multiple browser windows are opened.

This change improves the search logic by prioritzing windows to which
the sender window is anchored, before falling back to the last window
in the list.

Bug: 41482206
Change-Id: I4907abaac044953d7a5c2a5a5db54d357369b353

Validate-Test-Flakiness: skip
Change-Id: I4907abaac044953d7a5c2a5a5db54d357369b353
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5735992
Commit-Queue: Sohail Rajdev <sorajdev@microsoft.com>
Reviewed-by: Emilia Paz <emiliapaz@chromium.org>
Reviewed-by: Devlin Cronin <rdevlin.cronin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1334667}