[ios] Fix Profile report generation to work with name based API
On iOS, we want to refer to profile by name, not path, so change
ProfileReportGeneratorIOS to use GetBrowserStateByName(...). As
the underlying API only pass the path to the delegate, this needs
a workaround where the name is derived from the path. Eventually
this should be removed in favor of passing the name to this API.

Convert the rest of the reporting API to pass the name from the
ChromeBrowserState's GetBrowserStateName(...) instead of from the
path's base name.

Bug: 333865629, 356050207
Change-Id: I684a637f89861afd61302cc3fe6b717c10bd476a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5743500
Reviewed-by: Owen Min <zmin@chromium.org>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1334677}