[zlib] Add missing extension for PMULL in ARM64 GCC
A regression was introduced in the macro TARGET_ARMV8_WITH_CRC for
GCC. In the past it was empty and it would relay on the passed
architecture to support the required extensions. After the change it
explicitly adds an attribute for functions calling PMULL that
states that they require the target armv8-a+crc.

But PMULL belongs to CRYPTO extension, that is not provided by default
in ARMv8-a. So build breaks. To fix it, add the "crypto" extension to
the macro setting the armv8-a target.

Bug: 40565911
Change-Id: I44f35813197801f0f8dfa032901864123db75468
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5746951
Reviewed-by: Adenilson Cavalcanti <cavalcantii@chromium.org>
Commit-Queue: José Dapena Paz <jdapena@igalia.com>
Cr-Commit-Position: refs/heads/main@{#1334683}