Fix the usage of SelfKeepAlive in ai_text_session.cc
Both AITextSession::Responder and AITextSession::StreamingResponder are
using SelfKeepAlive. But they don't initialize the SelfKeepAlive.
For AITextSession::Responder, we need to initialize the SelfKeepAlive to
keep the Responder alive. But StreamingResponder don't need to use the
SelfKeepAlive, because StreamingResponder can be kept alive by the
ReadableStream which is the return value of the promptStreaming()
method.

So this CL, changes Responder to correctly initialize the SelfKeepAlive,
and removes SelfKeepAlive from StreamingResponder.

Bug: 325573146
Change-Id: Ice921e9020badf1e4d6be5a950292c6a3e3e018b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5748230
Reviewed-by: Mingyu Lei <leimy@chromium.org>
Commit-Queue: Tsuyoshi Horo <horo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1334710}